refactor: 유효성 검증을 삭제로직  안에서 수행하도록 변경

+++ src/main/java/qna/domain/Question.java
@@ -106,7 +106,9 @@ public class Question extends AbstractEntity {
         }
     }
 
-    public List<DeleteHistory> delete() {
+    public List<DeleteHistory> delete(User loginUser) throws CannotDeleteException {
+        validateDeletable(loginUser);
+
         deleted = true;
         List<DeleteHistory> deleteHistories = new ArrayList<>();
         deleteHistories.add(new DeleteHistory(ContentType.QUESTION, getId(), writer, LocalDateTime.now()));
